Restructured filesystem.
